Refactor authentication
