+++ src/application/service/authentication/Authenticate.ts
@@ -1,12 +1,11 @@
 import { LoginCommand } from '../../../presentation/commands/authentication/LoginCommand';
 import { AuthenticationMethod } from '../../../domain/model/user/AuthenticationMethod';
 import { AuthenticatorFactory } from '../../../domain/model/authentication/AuthenticatorFactory';
-import { AuthenticationIdentifier } from '../../../domain/model/user/AuthenticationIdentifier';
 
 export class Authenticate {
   constructor(private authenticatorFactory: AuthenticatorFactory) {}
 
-  public async execute({ method, identifier }: LoginCommand) {
+  public async execute({ method, authCode }: LoginCommand) {
     let authenticationMethod;
     try {
       authenticationMethod = AuthenticationMethod.fromString(method);
@@ -15,8 +14,7 @@ export class Authenticate {
     }
 
     const authenticator = this.authenticatorFactory.authenticatorFor(authenticationMethod);
-    const authenticationIdentifier = new AuthenticationIdentifier(identifier);
-    const token = await authenticator.authenticate(authenticationIdentifier);
+    const token = await authenticator.authenticate(authCode);
 
     return token;
   }
