+++ src/presentation/commands/authentication/LoginCommand.ts
@@ -1,4 +1,4 @@
 export interface LoginCommand {
   method: 'MAGIC_LINK' | 'ESTONIAN_ID';
-  identifier: string;
+  authCode: string;
 }
